Fix lwAFTR test to isolate changes between sub-tests
Python tests that share setup and teardown are "vulnerable" to side
effects from one subtest affecting future subtests.  In this case, the
set of tests resulted in multiple softwires terminating at ::1, which
then caused problems when projecting the binding table onto the
ietf-softwire model: whereas before there was no issue because the
answer that we were looking for happened to be first (or last?) in the
set of softwires associated with ::1, the change in hash function
changed the retrieval order, and thus caused the test to fail.

Hacked around by changing the tests to not define multiple softwires
with the same IPv6 B4 address.