Fix #835 - Support YANG choice statement
This adds support for YANG's choice statement (RFC6020 Section 7.9)
which allows for validation when two blocks can be specified exclusively.

The choice block is unique in a few ways and requires the introduction of
a new grammar handler. The choice statement is only there for configuration
purposes and should not appear itself in the data so it requires some
changes allowing for the defined "case" data to essencially replace the
choice statement.