Fix some small problems in YANG schema
This commit largely just cleans up a couple of issues. Most of
these are cosmetic such as fixing indentation etc. There were
a few typo cleanups and a missing curly brace added.

This also renames the resolved mac field to 'resolved-mac' from
'mac' to remain complient with the specification. Choice fields
do not allow for two fields to in different case blocks to have
the same name.

This commit also extends and clarifies the description on the
"device" field for the PCI address of the interface to configure.