Migrate lwaftr testing config files
This migrates the test data (config files) to the latest iteration
of the snabb-softwire-v2 schema. These changes most notably are the
addition of the "instance" and "queue" lists with the splitting of
the interface containers into instance spcecific and global.

The migration now also requires an option passing to the configuration
to be able to successfully migrate. This option is:

	/softwire-config/instance[device=PCI DEVICE]

A second optional option which when supplied configures the lwaftr
in the traditional/regular way of having a interface for interal (ipv6)
traffic and an interface for the external (ipv4) traffic. If not supplied
the lwaftr will be configured in on-a-stick mode. The option is:

	/softwire-config/instance/queue/external-interface[device=PCI DEVICE]

These options are provided with the new "-o" or "--option" flag with the
argument being these paths but with "PCI DEVICE" replace with the device.