Add support for snabb-softwire-v3 for reconfigurable
This also adds support for multiple arugments passed to the app when
one of the arguments ia yang schema. Prior to this it would fail to
properly encode app arguments if they were a table of arguments rather
than purely just a schema. It now ensures arguments are properly
encoded and decoded allowing for multiple.