ctable: Automatically detect need for scalar hash
When the key is a scalar type, the Lua representation of the key is
just a value and not a pointer to memory.  This is mostly transparent,
except for the hash function.  Before, the user had to override the
default hash function; now we do it on their behalf.

Additionally, prepare for ctables embedding a random key, and rename
"stride" in the lookup streamer to "width", to match other parts of
the code.