lib.pmu: Enable RDPMC via /sys/devices/cpu/rdpmc
We want to access the CPU PMU hardware directly instead of via Linux
system calls. This requires the kernel to allow access to the RDPMC
instruction from userspace processes (including Snabb which the kernel
does not know is accessing the PMU.)

This requires the /sys/devices/cpu/rdpmc value to be 2 on newer
kernels (4.0) or 1 on older kernels.

This change makes Snabb attempt to raise the /sys/devices/cpu/rdpmc
value up to the highest value that the kernel supports.