Add VM generated code to the repository
This has implications:

If you do not modify the VM then the generated code can be reused from
the source distribution. So: Just building the standard RaptorJIT from
source has fewer build dependencies because it does not need to
generate the VM code using DynASM (which depends on LuaJIT), etc.

If you hack the VM then you will need to check in the updated version
of the VM sources. Sorry about that. You will also need to put up with
merge conflicts on the generated code.

This may not be the best implementation but it is in the spirit of
reducing build dependencies when you are not actually hacking the VM:
https://github.com/raptorjit/raptorjit/issues/58#issuecomment-289789193