raptorjit.nix: preBuild = 'make bootstrapclean'
Start with 'make bootstrapclean' in builds under nix. This removes the
generated code that lives in-tree in the repository to ensure that it
will be rebuilt. This means that the nix build is not dependent on the
in-tree code being up-to-date.

Nix provides all of the build dependencies so there is no issue with
rebuilding the generated code.