Makefile: New 'reusevm' target & bootstrap-skipping method
This is a "less magic" and "more helpful" way to handle bootstrapping
vs not bootstrapping the VM.

The default behavior of "make" is reverted to its original behavior.
i.e. it generates code for the VM using DynASM (requiring LuaJIT.)

If LuaJIT is not available then this error message is printed:

    Error: Missing dependency (luajit) for bootstrapping the VM.

    Here are your options to build RaptorJIT:
      make reusevm     # copy reference VM from reusevm/
      (install luajit) # satisfy the dependency
      nix-build        # use nix to satisfy the dependencies

    Warning:
      Only reusevm when running a pristine copy of RaptorJIT from
      a repository that keeps the reference VM up-to-date.
      (Otherwise you might mix some stale code into your build.)

So then the user can choose between installing LuaJIT or running 'make
reusevm' to use the in-tree generated code at their own risk.

In practice it should be safe to 'reusevm' when building a pristine
master branch because CI is ensuring that the code is always
up-to-date. However, other branches or working copies may have stale
code in reusevm/ that is not safe to use.