Add checks for PCI devices and feed single instance configs to lwaftr
This adds proper checks that the PCI both exists and has an affinity for a numa
node.

It also now creates several versions of the configuration, each with only
one instance in the instance list and passes the config alone to the lwaftr app.
The lwaftr app then no longer needs the device passed in, it can get it by
extracting the only instance listed.