l2vpn: support dynamic reconfiguration
The l2vpn program supports a crude form of dynamic reconfiguration,
where the engine is run for a selectable interval after which the
configuration file is checked for changes.  If changes are detected,
the app network is reconfigured before the engine is restarted.

This kind of reconfiguraion is rather disruptive and it is probably
preferrable to restart the entire program when the configuration
changes until a smoother mechanism is developped.