vmprofile: Extend Lua API and add test cases
Extend the "Lua C API" for vmprofile to support allocating multiple
profiles and switching between them. This makes it easier to use.

Previously the expectation was that this functionality would be
implemented in Lua code using ljsyscall to allocate file-backed shared
memory for profiles and then the FFI to switch between them. (This is
still possible, too, and it works the same as it did before.)

Added test coverage to the test suite.