core.app: add a post_link() app method
The existing link() method is called for all add/remove actions per
link. In some cases, it is convenient or even necessary to let an app
perform actions after link configuration is completed but before the
app is run.  The method is only invoked when at least one link was
changed (added or removed).