apps.vlan.vlan: remove assert in VlanMux link()
The new semantics of the link() method no longer allow this kind of
check, because it is called for each link action separately and there
is no way of knowing when all links have been configured.  The
asserttions for the existence of the trunk port is moved to push().

Also remove backpressure in the forwarding loops.