lj_jit.h: Save 'parent' and 'exitno' fields in GCtrace
For debugging purposes it is very useful to be able to refer to the
origin of a trace (parent/exit) and so this change stores that
information persistently in GCtrace instead of only ephemerally in
jit_State.

These fields are now duplicated in jit_State (valid while recording)
and GCtrace (valid after recording.) This duplication could be avoided
by putting them only in GCtrace and accessing them via J->cur but this
would be a more noisy change since the existing fields are accessed
from many places including DynASM macros.