Restore --v4 and --v6 flags to `lwaftr run` command.
This restores the `--v4` and `--v6` option to the `snabb lwaftr run` command
which will perform an in-memory migration of the given configuration to the
PCI address(es) specified in the flags. This later should be expanded to
support listing multiple devices and for it to select those instances from
the configuration to run (excluding any others). See:

https://github.com/Igalia/snabb/pull/821#discussion_r130376854

As part of this change I've removed the now redundent migration and moved
the useful part of the code to `run.lua`. There is no real need for there
to be two ways. The tests have been updated to simply use the in-memory
migration the `run` command now provides. For persistant changes users
can modify the config file by hand (as with any other config changes).