lj_trace.c: Make trace numbers unique (stop reusing them)
Make JIT trace numbers unique for completed traces. This way a trace
number is an unambiguous way to refer to a GCtrace object even across
flushes over the JIT.

The immediate motivation is to make the vmprofile data valid across
JIT flushes. It is counting samples by trace number and so when trace
numbers are reused then multiple traces can "collide" on the same
profiler counter.