Fix #876 - Improve config migration by chaining together
Each migration is now only responsible for progressing to the next version
which then will run the subsequent migration to progress further. This is
done as a long chain until the configuration has been migrated to the latest
configuration schema.

This also fixes an issue with the 3.2.0 -> 2017.07.01 migration which no
longer functioned correctly.