virtual_ether_mux: Treat vlan=0 as vlan=nil
Commit da4f98b09caf015e800fb23b8ea0ef4a27bea85a makes apps.vlan
validate the vlan ID argument. This correctly rejects the value zero
as invalid (not allowed on the wire in 802.1Q standard.) However, the
snabbnfv application does permit vlan=0 in its configuration file and
this is used in various test vector files. This causes test failures.

There are two obvious solutions. The first is to update all
configurations to never use the invalid vlan=0. The second is to
interpret vlan=0 as meaning "no VLAN." This commit implements the
second one.