Deprecate reconfigurable flag
This deprecates the --reconfigurable flag from all lwaftr apps. The
reconfigurable / multiprocess mode is always enabled. The tests have
been simplied to reflect this. A deprecation warning is shown if
--reconfigurable is used.

As of this change, ingress drop monitor is no longer supported. An
issue tracking the support is #902. A warning is shown if it is used.