Start worker processes from leader
This commit changes the setup of the lwAFTR to run the worker
processes from the Leader process instead of from the setup
itself.

There is a new argument for the leader which is the function to start
the worker (follower). The worker function needs to setup the follower
and return the PID of it.

There is also a change to the setup function that requires it to
return a table of ID to app graph. The setup function should return
the same ID for the same function (i.e. it should not be random but
consistantly unique to itself).