Maintain a single in_place_dependencies
This change makes it so that the leader maintains a single set
of dependencies rather than each follower maintaining their own. To
do this the table backing each dependency also now lists a table of
follower IDs so that it's known which followers need to restart.

The main change to this is that the following support functions now
take a table with id and app graph rather than simply the app graph:

- update_mutable_objects_embedded_in_app_initargs
- compute_apps_to_restart_after_configuration_update