Stabilize breathe order
Since the switch to use a sorted breathe order, we have been running
apps only once per breath.  However the particular computed sort order
could vary from run to run if there are strongly connected components in
the graph, as can happen with "middlebox"-style apps that sit in the
middle of a bidirectional graph, like the NDP app.  This commit
stabilizes the sort order, breaking ties by app name instead of Lua
table iteration order.  The hope is that variance in this kind of graph
will be reduced.