lj_trace.c: Clear penalty slot after successful trace
Clear the penalty slot associated with a bytecode after it is traced
successfully.

This prevents penalties from accumulating for bytecodes that are
traced frequently, such as branchy subroutines that are called from
many different root traces and will need to record a side-trace for
each.

Especially intended to handle the special case when applications are
generating code at runtime that creates a "fairly large" number of
root traces and that need correspondingly many side-traces to be
recorded without prematurely and unproductively blacklisting things.