GCtrace: Make it possible to map between IR and mcode insns
Add a 'szirmcode' array to GCtrace that keeps track of how many bytes
of machine code were assembled for each IR instruction.

This allows us to construct an association between IR code and mcode,
either to see the mcode that was assembled for a given IR instruction,
or to map a machine code instruction back to the relevant IR instruction.