Delegate exporting packets to FlowSets
FlowSets can generate the entire export packets by giving it access to
the parent IPFIX object that created them.  This avoids queueing on an
internal link and an additional loop over that queue.

The IP and UDP headers are pre-allocated in the constructor to avoid
allocations in the add_transport_headers() method.

Add options to set the source and destiantion MAC addresses of export
packets.