ctable: fix power-of-2 issue
LuaJIT dislikes pointer subtractions when the size of the pointed-to
objects is not a power of two, which leads to trace aborts. ctable
suffers from this problem. The patch pads the cdata type used as entry
for the table to the next power of two.

We also disable tail-call optimization for functions that end with a
call to a built-in, which also causes trace aborts in certain cases.