Allow multiple files as input for make_ipfix_element_map()
The main purpose of this is to be able to store non-standard element
definitions in a separate file, like that for the reduced-size
encoding of tcpControlBits, which is also part of this commit. The v4
and v6 templates now use this instead of tcpControlBits to be
compatible with current versions of the nfdump suite.