Support VLAN-tagged traffic, IPv6 extension headers
The current version of pflua cannot handle VLAN tags at all and IPv6
extenstion headers only in a very limited fashion (the latter is
mostly a problem with the underlying pflang language).

To work around these issues as well as to simplify address-family
independent code in the extractor and accumulator methods of
templates, all packets are pre-processed before they are passed to a
template's matching function.

The pre-processor recognizes 802.1q-tagged packets and arranges that
the matching function sees the proper Ethertype.  It also detects
whether a packet is either unfragmented or is the initial fragment of
a fragmented packet (in which case the payload does not contain a L4
header).  It also detects all IPv6 extenstion headers that might be
present and removes them from the packet such that the matching
function always sees a plain IPv6 packet with the fixed header only.

Information about the VLAN tag, fragmentation status, location of L3
and L4 headers in the packet as well as whether the packet is
truncated or not (e.g. when received from a port-mirror that only
copies the beginning of a packet) is added as a "metadata block" just
after the packet data for access by the extractor/accumulator methods.