Support starting and stopping instances via snabb config
This commit supports the addition of new instances by snabb config add
which will start a new follower and configure it with the instance
provided. Snabb config remove on an instance will now also stop the
instance.

Tests have been added to verify it starts and stops followers as
described above.