Update ietf-softwire translation layer for multiproc
This updates the ietf-softwire translation layer to support the
multiprocess schema changes. It can now get multiple instances and
remove an instance.

Set cannot occur as there isn't enough information to add a new
instance  (such as the PCI addresses of the interfaces). This is the
same reason why add cannot be used either.