Standardize on "app:rxdrop()" instead of "app:ingress_packet_drops()"
The ingress drop monitor used to poll apps with an
"ingress_packet_drops" function, but the intel_mp app names this
function "rxdrop".  "rxdrop" is much more succinct and captures the
meaning, so let's use that instead.  Change intel10g to use the new
"rxdrop" name, and update the lwaftr loadtest.