Stop csv stats outputting for stale instances
This performs a similiar check to what the leader's doing when
shutting down followers. It'll check if the follower it's responsible
for is still alive via `waitpid` and if not it'll disable the timer.

This is to prevent the CSV stats from printing indefinitely for
instances which no longer exist.