Add test for migration using --on-a-stick command
When --on-a-stick command is used with a single instance defined it
should migrate the instance to work in a stick  (replace the device
with the one provided in the flag and remove the device in
`external-device`). It just checks that it works by not crashing and
says it's going to do it in the output.