Appropriately choose on-a-stick or bump-in-the-wire
Previously you could either configure the lwaftr with --v4/--v6 and
--on-a-stick and it'd then migrate the configuration provided and then
on every instance added subsequently it'd only use the mode selected
at the start despite what the configuration says. If no flag was
specified it'd only use --v4/--v6 for everything including new
configurations added without two devices.

It will now select the correct method based on the configuration
except in the case of --vertio which will cause it to always use the
virtual driver. Configurations can now contain a mix between
bump-in-the-wire and on-a-stick configurations and it'll choose it
appropriately. New configurations added by the `snabb config` tools
will also choose it based on the instance configuration for each
instance.