Adjust how intel_mp VMDq enabled state is checked
This is in preparation for changing how the MRQC register
is set for different queuing modes for the 82599. We don't
want to rely on this register state to check if VMDq is on.

Instead, use shm to communicate between processes. This
should also be used for VMDq pool coordination too.