Adjust how VMDq pool selection works
This approach is less hacky (uses the NIC's semaphore
register instead of a file lock) and reuses the shared
memory directory that is needed for coordinating VMDq
enabled status.

Note that the shared memory needs to be initialized
by the main process every time since we can't rely
on stop() cleanup.