Add alarms API method to set default alarms
These list of default alarms define a values that will always
overwritten when adding an alarm that matches the default.

The matching is done only by alarm_type_id and alarm_type_qualifier
since resource is meant to be a dinamically defined value and it's not
possible to know it beforehand.