Fix insertion into sparse cltables
Previously we would do a table.insert, then assume the index was
inserted at #values.  However the # operator just returns a filled
index for which the next index is nil, and thus table.insert could
insert into the middle of a table but we don't know afterwards where
that insertion was!  So, instead use # to find an insertion point, and
don't use table.insert.