Change default schema to ietf-softwire
This updates the leader so you can specify a 'default-schema' and have
it be the default for the `--schema` flag on the `snabb config`
commands. If this is not provided like usual it falls back to whatever
the native schema is.

The lwaftr's tests have been updated to explicitly always provide
the schema flag.