lj_record.c: Record IFUNC/IFUNCV the same as FUNC/FUNCV
Record calls to blacklisted functions the same way as calls to
uncompiled functions. This allows for the possibility that a
blacklisted function can be successfully recorded as part of a larger
trace.

This is specifically intended to fix the problem where a tiny utility
function somewhere, such as `ntohs`, is not able to be compiled in
isolation for some reason and that ends up preventing all of its
callers from compiling too.

This has tended to happen for obscure reasons like "the utility
function ends with a tail-call to a C library routine" and with this
patch we should not need to worry about such minutiae anymore. (We
should be able to retire the "wrap in parenthesis to prevent a
tailcall" trick that some people are familiar with.)

More background:

- How the trace recorder treats function calls: raptorjit/raptorjit#118.
- Why this can be a problem: raptorjit/raptorjit#119.
- More on this specific solution: raptorjit/raptorjit#120.