Attempt to fix apps/intel_mp/test_1g_sw_sem.snabb
I think the original intent of this test was to test intel_mp without
any rxq/txq, as in test_10g_sw_sem.snabb.  However in the refactoring of
the default values of rxq/txq and how lib.parse was used to provide
defaults, rxq/txq got turned on, which prevents the test from working as
nic:stop recursively acquires the semaphore.