vmprofile 3.0 with more information
This is an extension to the data collected by vmprofile and its file format:

The overall vm state counters are extended with the new fine-grained
definitions introduced in 089c55a that also covers JIT mode operation.
Now the profiler will always bump exactly one VM state counter. So if
you don't care about per-trace information you don't have to look at
it, since the overall summary is complete in itself, and the per-trace
information is only a supplementary breakdown.

Traces now count 'interp' time, which is the time spent in the
interpreter due to an exit at the end of that trace. This is for
working out which trace to "blame" when the interpreter is hot.

The 'other' counter is renamed to 'ffi' because in practice we always
assume that is the reason.