Remove feature to compress GCproto.lineinfo
The source code line number of each bytecode is now always stored in
memory as a 32-bit number.

Previously a simple compression scheme was used: an 8-bit number was
used for <256 line functions, a 16-bit number for <64K line functions,
and a 32-bit number otherwise.

This specific optimization is not appropraite for RaptorJIT. On the
one hand there will never be a practical benefit on any server
platform because the amount of memory involved is miniscule, and on
the other hand it complicates the JIT data structures that need to be
understood by new maintainers and also supported in external tools
like Studio.

(Studio needs to understand the RaptorJIT data structures and in this
case I decided to do that by removing code from RaptorJIT rather than
adding code to Studio.)