VMProfile 4.0: File format revision
Now data is stored in a 2D matrix indexed by VM state and trace number
(0 for "other.") This holds for all VM states including interpreting,
recording, etc.

This makes the format simpler and more regular. Previously the samples
were stored in two related sections, one global and one per-trace. Now
it is all per-trace with 0 as a catch all. (Just sum the values for
all traces to compute the total "global" values.)