intel_mp: add queue enable counters, fix rxdrop counter
Two counters are added for each queue called "q<i>_<rx|tx>_enabled",
which is set to one if the corresponding queue is enabled and zero
otherwise.  This helps consumers of the statistics to determine which
of the per-queue counters contain valid data.

Queues are enabled implicitly when an instance of the driver is
initialized but the counters are maintained by the master process
only.  The only way for the master to detect the state of a queue is
by an explicite check in sync_stats().

For each process, two new counters txq and rxq are created which hold
the number of the transmit and receive queues for which the driver was
configured.

For the 82599, receive drop counters are maintained per rx queue.  The
total rx drop counter must be derived from the per-queue counters.
This commit assumes that each enabled receive queue uses its "natural"
statistics register.