lj_auditlog.c: Add lj_auditlog_open(path) API function
Open the auditlog on the filesystem. The log will include all entries
since the process started i.e. it includes events that happened before
the log was opened. This is important so that the log will always be
complete. (Log entries are buffered in RAM, up to a limit, before the
first log file is opened.)