Rework alarms timestamping logic
The conversion of localtime to ISO8601 format didn't include timezone
information. For instance, timestamp 0 (UNIX epoch) was converted to in
GMT +01:00:

1970-01-01T00:00:00Z

While it should be:

1970-01-01T00:00:00Z+01:00

So now all alarms date information works using localtime and converting
it to ISO8601 including timezone info.  When an ISO8601 date is
converted to seconds it takes into account timezone information. That
makes possible to compare local timestamps and ISO8601 dates.