lj_parse.c, lj_debug.c: Remove clever tricks in lineinfo encoding
Remove two clever tricks whose benefits don't justify their complexity:

The GCproto.lineinfo table used to exclude the first and last
bytecodes. The line numbers of those bytecodes had to be calculated as
a special case (using other state in the GCproto struct.) Now every
bytecode has an entry in the lineinfo table and is treated the same.

The GCproto.lineinfo table used to store delta values relative to the
first line. Now the absolute line number is stored. (The deltas helped
with a compression scheme that has previously been removed from RaptorJIT.)