Support "one-liner" usage of jit.vmprofile.open()
Make simple usage of VMProfile more convenient.

For example:

    jit.vmprofile.open("bigloop")   for i = 1, 1e9 do end
    jit.vmprofile.open("shortloop") for i = 1, 1e6 do end

This involves two changes to the jit.vmprofile API:

Preload the jit.vmprofile library, so that no require() is needed.

Make the default behavior of jit.vmprofile.open(filename) include
selecting the profile and enabling the timer. This way a single call
with a filename will enable live profiling to disk.

The slightly awkward aspect is that the jit.vmprofile.open() changes
are not quite backwards compatible. The old behavior was to not select
the new profile nor enable the timer (on the assumption that this will
be done separately.) This behavior can be accessed using new optional
arguments boolean arguments `noselect` and `noopen`:

    jit.vmprofile(filename, [noselect, noopen])