core.link: free any packets left in link when freeing link
Due to how the engine works, the case that a link is non-empty between
breaths is rare (specifically, it should only occur when an app pushes
to a “slow” output interface). That’s probably why this went unnoticed
until now. I don’t even have a case that triggers this bug, it just came
to me.