Add DNS-SD program
DNS-SD (DNS Service Discovery) is the protocol used by Apple's Bonjour
to discover services and devices in a local network.  It's widely used
as well for discovering IoT devices such as Chromecast, Amazon FireTV or
services such as Spotify in a local area network.

Snabb's dnssd program is similar to avahi-browser. Given an OS network
interface sends a multicast query to discover the services and devices
available in the network.

The program reads incoming traffic and prints out responses.  Currently
supported DNS records are PTR, A, SRV and TXT.