Just use the default gcc instead of clang 4.0.0.
It was a bit of a flight of fancy to use a pinned version of clang.
This would make sense if we were really depending on hyper-specific
behaviour of the C compiler, like some automatic SIMD parallization,
but in reality we're not. So don't be fussy and just keep it simple.