Allow multiple IPFIX instances in the same process
This commit prepares for the use case where an RSS app is feeding
multiple IPFIX instances running in the same process as the RSS app
itself.  Those instances are distinguished by a numeric "instance id",
which is is attached to the names of the apps to make them unique.
The instance id is also used to group template counters per instance.

If run in combination of the RSS app, packets already carry meta-data
when passed to IPFIX.  A configuration option has been added to
disable adding of packet meta-data by IPFIX (default is enabled).