Refactor template definitions
Templates are now defined as elements of a table to be able to refer
to them by name for configuration.

The names of the templates passed to the IPFIX constructor can now
have an optional parameter attached to it which sets a
template-specific cache size. The size is appended to the name of the
template separared by ":'.