Sizes for "packetblaster lwaftr" are frame sizes
This change makes it so that the sizes passed to "packetblaster lwaftr"
are frame sizes instead of packet sizes without the ethernet headers.

This seems to better correspond to the intent of including 64 as a
default size.  Often you want to generate test traffic of the minimum
frame size, and this makes that goal more attainable.

Additionally this commit fixes a bug whereby the frame size was longer
than the actual packet data, because the length in the UDP headers was
wrong.