core.app: remove garbage when using pace_breathing
The accumulators lastfrees, lastfreebytes, lastfreebits for the
counters frees, freebytes and freebits are initialized as Lua numbers
but implicitly converted to cdata objects in the assignments in
pace_breathing(). This causes allocations that cannot be removed by
the sink optimizer.  Conversion to Lua numbers avoids this and reduces
GC noise.