lib.interlink: fix life-time fsm design and implementation bugs
As it happens, quite a few edge (and not so edge) cases in the state machine
that manages an interlink’s life-time were previously overlooked due to overly
narrow testing. Prior to this commit, multiple concurrent receivers /
transmitters were not handled properly, among other things. This patch fleshes
out the state machine in a way that should hopefully live up to the
requirements.

Externally visible changes include the split of new and free to attach_receiver
/ attach_transmitter and detach_receiver / detach_transmitter. This is
obviously necessary, since lib.interlink has to ensure that an interlink is
never used by two receivers or two transmitters at the same time.

It should now also possible to “reuse” interlink queues. I.e., it is possible
for processes to detach and reattach without reallocating the queue. This is
handy for scenarios in which interlinks are used for hot fail-over, and basic
use cases like restarting group process individually without restarting the
process at the other end of the queues.

In the future, we should really add a comprehensive test suite for the
interlink apps that exercise all these cases!