lib.interlink / core.main: fix shutdown bugs
59caa02ca (apps.interlink: shutdown handler to release stale attachments)
introduced a bug where the supervisor processes would create directories under
SNABB_SHM_ROOT. This happened because lib.interlink depended on core.link and
core.packet, and I mistakenly added a call to initialize() to main in order to
resolve these dependencies.

The dependencies existed because detaching possibly involves freeing any
packets remaining on the queue, and the queue size was based on link.max (for
no particular reason.)

Besides introducing this bug, this approach was wrong for another reason: the
supervisor has no packet freelist, and hence there is no point in freeing
packets within the shutdown code.

This commit removes the unnecessary dependency on core.link, and makes sure
that the dependency on core.packet is only exercised when it makes sense to do
so. E.g., when detaching from queues in the supervisor no freeing of packets
takes place. Finally, the call to initialize() in the supervisor is reverted.