lj_trace.c: Allow reuse of trace numbers after jit.flush()
RaptorJIT had earlier adopted the behavior of keeping trace numbers
unique between JIT flushes (up to the limit of 4096 beyond which
numbers would be reused.) This was intended to prevent two different
traces from sharing a profiler bucket.

However, this is a primitive solution, and we now intend to extend the
profiler to explicitly track multiple generations of traces.