core.packet: enable group freelist only on demand
This adds a function enable_group_freelist() that will enable the group
freelist, its idempotent. The interlink receiver and transmitter apps call it
during initialization to ensure the group freelist is present/enabled.

Due to JuJITsu the extra branch to check whether the group freelist is enabled
is expected to be virtually free.