dasm: Error when a 64-bit value is used as a 32-bit immediate
If a 64-bit value is provided as a 32-bit immediate operand for a DynASM
instruction then error. The previous behavior was to automatically
truncate to 32-bit.

This is particularly significant for GC64 mode where pointers to
objects created by `ffi.new()` cannot be safely truncated to 32 bits.

The solution for referencing 64-bit immediates is to use 'mov64' and
this is suggested in the error message.