Add basic IO streams
The goal is to replace Lua's "io" library with a compatible interface,
except written in Lua instead of C, and allowing extra functionality
such as string files, mmap buffers, and non-blocking concurrency.

This commit adds the common stream functionality, which corresponds to a
superset of the interface of Lua files, as implemented by LuaJIT.  Note
that some functionality is present but not implemented.