lj_state.c: Partial fix of assertion on shutdown
Partially fix deallocation of state when the VM state is freed.

This is not correct, and leaves some memory allocated, and I have
temporarily disabled the assertion that this breaks. Just because
being able to run with assertions is currently more important than
being sure that all memory is freed at closing.