YANG syntax parser is now streaming
This commit makes the YANG parser accept a stream instead of a string,
and adapts all eventual callers to work in streams instead of strings,
creating temporary streams as needed.

Eventually the RPC protocol between the ptree manager and the "snabb
config" client should also be changed to allow streaming (not requiring
the config to be written out as a string).